Add files via upload
This project is a data analysis and machine learning project for predicting house prices in King County, USA. The dataset used in this project contains information on the sale of houses between May 2014 and May 2015.

The project begins with data cleaning and exploration, including handling missing values and outliers, checking for multicollinearity, and visualizing relationships between variables.

Next, a linear regression model is fitted to the data using 11 features related to the houses, and the model's performance is evaluated using R-squared metric.

Then, a Ridge regression model is fitted to the data using the same set of features but with regularization parameter set to 0.1. The model's performance is evaluated using R-squared metric as well.

Finally, a pipeline is created to preprocess the data and fit a polynomial regression model to the data. The pipeline includes scaling the features, generating polynomial features, and fitting a linear regression model. The model's performance is also evaluated using R-squared metric.

Overall, this project demonstrates data cleaning, exploration, and preparation techniques, as well as machine learning modeling and evaluation techniques. The project aims to provide insights into the factors that affect house prices and to develop accurate models for predicting house prices based on the given features.